Seperate pokemon class
