+++ Pokemon/Player.cpp
@@ -1,6 +1,7 @@
 #include "Player.hpp"
 #include "PokemonType.hpp"
 #include "PokemonChoice.hpp"
+#include "Pokemon.hpp"
 #include "Utility.hpp"
 
 #include <iostream>
