+++ Pokemon/Player.hpp
@@ -1,5 +1,5 @@
 
-
+#include "Pokemon.hpp"
 #include <string>
 using namespace std;
